track-weekend-gold-price-tokenized-commodities/requirements.txt
requests>=2.31.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.20.0
python-dotenv>=1.0.0
